+++ my-diary/src/routes/Router.jsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Route, Routes, Navigate } from 'react-router-dom'
 import { useAuthContext } from '../hooks/useAuthContext'
 import Nav from '../components/Navbar/Nav'
 import Home from '../pages/home/Home'
-import Login from '../pages/account/Login'
+import Login from '../pages/account/Login/Login'
 import Signup from '../pages/account/Signup'
 import Off from '../pages/etc/Off'
 import Footer from '../components/Footer/Footer'
